+++ src/lang/nodes/Postfix.js
@@ -1,26 +1,18 @@
 export function compile(node, ctx) {
     if (node.operator && node.operator.type) {
+        ctx.put('($=>(');
         ctx.createScope(
-            () => {
-                ctx.put('($=>(');
-                ctx.node(node.operator);
-                ctx.put('))');
-            },
-            (scopeStart, sp) => {
-                return scopeStart + sp + ',';
-            }
+            () => ctx.node(node.operator),
+            (sp) => sp + ','
         );
+        ctx.put('))');
         ctx.put('(');
         ctx.node(node.argument);
         ctx.put(')');
         return;
     }
 
-    switch (node.operator) {
-        default: {
-            ctx.error('Unknown operator "' + node.operator + '"', node);
-        }
-    }
+    ctx.error('Unknown operator "' + node.operator + '"', node);
 }
 export function walk(node, ctx) {
     ctx.node(node.argument);