+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## next
 
-- Improved suggestions for the ternary operator (`?:`)
 - Disallowed duplicate parameter names in function definitions, i.e., `($a, $a) => expr` is now prohibited
+- Improved suggestions for the ternary operator (`?:`)
+- Fixed suggestions for `and`, `or` and `??` operators
 - Fixed a compilation error for the generated code of `function` expressions in stat mode in some cases
 - Fixed a runtime error for a `comparator function` in some scenarios
 - Fixed support for TypedArrays with slice notation and methods such as `numbers()`, `avg()`, `count()`, `sum()`, `median()`, `variance()`, `stdev()` and `percentile()`